+++ packages/file-picker/android/src/main/java/io/capawesome/capacitorjs/plugins/filepicker/FilePicker.java
@@ -18,17 +18,24 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.Objects;
 
 public class FilePicker {
 
     public static final String TAG = "FilePicker";
-    private FilePickerPlugin plugin;
+    private final FilePickerPlugin plugin;
 
     FilePicker(FilePickerPlugin plugin) {
         this.plugin = plugin;
     }
 
-    public void copyFile(@NonNull Uri from, @NonNull Uri to) throws Exception {
+    public void copyFile(@NonNull Uri from, @NonNull Uri to, Boolean shouldOverwrite) throws Exception {
+        if (!shouldOverwrite) {
+            File file = new File(Objects.requireNonNull(to.getPath()));
+            if (!file.exists()) {
+                throw new Exception(FilePickerPlugin.ERROR_FILE_ALREADY_EXISTS);
+            }
+        }
         InputStream inputStream = plugin.getBridge().getContext().getContentResolver().openInputStream(from);
         OutputStream outputStream = plugin.getBridge().getContext().getContentResolver().openOutputStream(to);
         if (inputStream == null || outputStream == null) {