+++ packages/libsql/android/src/main/java/io/capawesome/capacitorjs/plugins/libsql/interfaces/Result.java
@@ -1,0 +1,7 @@
+package io.capawesome.capacitorjs.plugins.libsql.interfaces;
+
+import com.getcapacitor.JSObject;
+
+public interface Result {
+    JSObject toJSObject();
+}